+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - new parameter called `allow-archived-tags` to allow selection of archived tags.
+- new flag `tag` on `report` commands, this will filter the time entries with all the tags informed.
 
 ## [v0.40.0] - 2022-08-09